requirements.txt
beautifulsoup4>=4.11.1
pymysql>=1.0.2
google-api-python-client>=2.65.0
gspread>=5.7.0
oauth2client==4.1.3
openpyxl>=3.0.10
pymongo>=4.3.1
requests>=2.28.1
configparser>=5.3.0
pandas>=1.5.1